+++ legacy/txstore/json.go
