+++ legacy/txstore/serialization.go
